speed up smiulation by not sorting vectors of simulated distances
We can track of the number of sinmulations in which the simulated value
is smaller than the observed distance. This requires comparing the
observed value to every simulated value (O(n), but this should be faster
than sorting the simulated distance vectors (O(n*log(n)), and integrating
them with a merge sort, then bisecting to find the position of the
observed distance.